+++ lib/utils/common.py
@@ -24,7 +24,6 @@ from urllib.parse import quote, urljoin
 
 from lib.core.settings import (
     INVALID_CHARS_FOR_WINDOWS_FILENAME,
-    INSECURE_CSV_CHARS,
     INVALID_FILENAME_CHAR_REPLACEMENT,
     IS_WINDOWS,
     URL_SAFE_CHARS,
@@ -68,11 +67,14 @@ def get_valid_filename(string):
     return string
 
 
-def human_size(num):
+def get_readable_size(num):
     base = 1024
-    for unit in ["B ", "KB", "MB", "GB"]:
+    units = ("B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB")
+
+    for unit in units:
         if -base < num < base:
             return f"{num}{unit}"
+
         num = round(num / base)
 
     return f"{num}TB"
@@ -92,14 +94,6 @@ def iprange(subnet):
         network = IPv6Network(subnet)
 
     return [str(ip) for ip in network]
-
-
-# Prevent CSV injection. Reference: https://www.exploit-db.com/exploits/49370
-def escape_csv(text):
-    if text.startswith(INSECURE_CSV_CHARS):
-        text = "'" + text
-
-    return text.replace('"', '""')
 
 
 # The browser direction behavior when you click on <a href="bar">link</a>