+++ lib/view/terminal.py
@@ -17,13 +17,11 @@
 #  Author: Mauro Soria
 
 import sys
-import time
 import shutil
 
 from lib.core.data import options
 from lib.core.decorators import locked
 from lib.core.settings import IS_WINDOWS
-from lib.utils.common import human_size
 from lib.view.colors import set_color, clean_color, disable_color
 
 if IS_WINDOWS:
@@ -88,21 +86,20 @@ class CLI:
             self.buffer += "\n"
 
     def status_report(self, response, full_url):
-        status = response.status
-        length = human_size(response.length)
         target = response.url if full_url else "/" + response.full_path
-        current_time = time.strftime("%H:%M:%S")
-        message = f"[{current_time}] {status} - {length.rjust(6, ' ')} - {target}"
+        # Get time from datetime string
+        time = response.datetime.split()[1]
+        message = f"[{time}] {response.status} - {response.size.rjust(6, ' ')} - {target}"
 
-        if status in (200, 201, 204):
+        if response.status in (200, 201, 204):
             message = set_color(message, fore="green")
-        elif status == 401:
+        elif response.status == 401:
             message = set_color(message, fore="yellow")
-        elif status == 403:
+        elif response.status == 403:
             message = set_color(message, fore="blue")
-        elif status in range(500, 600):
+        elif response.status in range(500, 600):
             message = set_color(message, fore="red")
-        elif status in range(300, 400):
+        elif response.status in range(300, 400):
             message = set_color(message, fore="cyan")
         else:
             message = set_color(message, fore="magenta")
@@ -197,9 +194,6 @@ class CLI:
         self.new_line()
         self.print_header({"Target": target})
 
-    def output_location(self, file):
-        self.new_line(f"\nOutput: {file}")
-
     def log_file(self, file):
         self.new_line(f"\nLog File: {file}")
 
@@ -224,9 +218,6 @@ class QuietCLI(CLI):
         pass
 
     def target(*args):
-        pass
-
-    def output_location(*args):
         pass
 
     def log_file(*args):