+++ tests/report/test_reports.py
@@ -22,24 +22,17 @@ from unittest import TestCase
 
 from lib.connection.requester import Requester
 from lib.core.settings import DUMMY_URL, DUMMY_WORD, NEW_LINE, TMP_PATH
-from lib.reports.csv_report import CSVReport
-from lib.reports.json_report import JSONReport
-from lib.reports.markdown_report import MarkdownReport
-from lib.reports.plain_text_report import PlainTextReport
-from lib.reports.simple_report import SimpleReport
-from lib.reports.xml_report import XMLReport
+from lib.report.json_report import JSONReport
+from lib.report.markdown_report import MarkdownReport
+from lib.report.plain_text_report import PlainTextReport
+from lib.report.simple_report import SimpleReport
+from lib.report.xml_report import XMLReport
 
 requester = Requester()
 test_entries = [requester.request(DUMMY_URL + DUMMY_WORD)]
 
 
 class TestReports(TestCase):
-    def test_csv_report(self):
-        expected_output = "URL,Status,Size,Content Type,Redirection" + NEW_LINE
-        expected_output += f"{DUMMY_URL}{DUMMY_WORD},404,648,text/html," + NEW_LINE
-
-        self.assertEqual(CSVReport(TMP_PATH).generate(test_entries), expected_output, "CSV report is unintended")
-
     def test_json_report(self):
         output = JSONReport(TMP_PATH).generate(test_entries)
         expected_results = [{
@@ -58,7 +51,7 @@ class TestReports(TestCase):
         self.assertTrue(MarkdownReport(TMP_PATH).generate(test_entries).endswith(expected_table))
 
     def test_plain_text_report(self):
-        expected_result = f"404   648B   {DUMMY_URL}{DUMMY_WORD}" + NEW_LINE
+        expected_result = f"404   648B {DUMMY_URL}{DUMMY_WORD}" + NEW_LINE
         self.assertTrue(PlainTextReport(TMP_PATH).generate(test_entries).endswith(expected_result))
 
     def test_simple_report(self):
@@ -69,5 +62,6 @@ class TestReports(TestCase):
         expected_result += "\t\t<status>404</status>\n"
         expected_result += "\t\t<contentLength>648</contentLength>\n"
         expected_result += "\t\t<contentType>text/html</contentType>\n"
+        expected_result += "\t\t<redirect/>\n"
         expected_result += "\t</target>"
         self.assertTrue(expected_result in XMLReport(TMP_PATH).generate(test_entries))